+++ packages/nostrify/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `@nostrify/react` package.
+- NPool: add `.group` method to create sub-pools.
+
 ## 0.41.2 - 2025-04-10
 
 ### Added